web-editor-backend/pyproject.toml
[project]
name = "web-editor-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "python-multipart>=0.0.9",
  "pydantic>=2.6",
  "python-docx>=1.1.0",
  "pymupdf>=1.23",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.26",
  "ruff>=0.4",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"app/main.py" = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
